pyproject.toml
[project]
name = "research-agora"
version = "1.0.0"
description = "Research Agora - A plugin marketplace for Claude Code"
requires-python = ">=3.9"

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-xdist>=3.0.0",
    "pyyaml>=6.0",
]
site = [
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
]
dev = [
    "pre-commit>=3.0.0",
]
synthetic = [
    "anthropic>=0.40.0",
    "playwright>=1.48.0",
    "pytest-asyncio>=0.24.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "-ra",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests requiring external resources",
    "synthetic: LLM-based synthetic user tests (requires ANTHROPIC_API_KEY)",
]
